pyproject.toml
[project]
name = "penetration-testing"
version = "0.1.0"
description = "Self-hosted penetration testing platform"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    # Task queue
    "celery[redis]>=5.4.0",
    "redis>=5.2.0",
    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "email-validator>=2.0.0",
    # Config
    "pydantic-settings>=2.6.0",
    # CLI
    "typer>=0.14.0",
    "rich>=13.9.0",
    # Scanning
    "sslyze>=6.1.0,<7.0.0",
    "python-nmap>=0.7.1",
    "httpx>=0.28.0",
    # Monitoring
    "flower>=2.0.0",
    # Rate limiting
    "slowapi>=0.1.9",
    # Docker SDK
    "docker>=7.0.0",
    # CVSS scoring
    "cvss>=3.0",
    # Report generation
    "weasyprint>=62.0",
    "jinja2>=3.1.0",
    # Celery encryption
    "cryptography>=42.0.0",
    # HIBP bloom filter
    "bitarray>=2.9.0",
    # Metasploit RPC
    "pymetasploit3>=1.0.0",
]

[project.scripts]
pentest = "src.cli.app:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.6",
]
